split admin into multiple files
