+++ posthog/admin/admins/instance_setting_admin.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+
+
+class InstanceSettingAdmin(admin.ModelAdmin):
+    list_display = (
+        "id",
+        "key",
+        "value",
+    )
