+++ posthog/admin/inlines/totp_device_inline.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from django_otp.plugins.otp_totp.models import TOTPDevice
+
+
+class TOTPDeviceInline(admin.TabularInline):
+    model = TOTPDevice
+    extra = 0
